Add receive-with-timeout IPC syscall
Introduce a receiveTimeout syscall and wire timeout support through the IPC path. Adds a new ReceiveTimeoutSyscall provider, an _asm_recv_timeout assembly entry, and a @_silgen_name binding. Extends SyscallNumber, RXIPC wrapper (receive(handle:timeout:)), and IPCInterface.receive to accept timeoutTicks. RendezvousIPC now sets/clears ipcDeadline and calculates deadlines when blocking; Process.ipcDeadline is widened to UInt64. Virtual timer handler now checks IPC timeouts each tick and only preempts when a quantum actually expired. These changes implement timed receive semantics and timeout bookkeeping in the kernel.

+++ Sources/ReixKernel/Process/Process.swift
@@ -50,7 +50,7 @@ public struct Process: RXEntry {
     
     public var message: Message?
     public var replyTo: UnsafeMutablePointer<Process>?
-    public var ipcDeadline: UInt32?
+    public var ipcDeadline: UInt64?
     public var expectsReply: Bool = false
     public var pendingGrant: UInt32? = nil
